+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ nameko-rediskn versions, where semantic versioning is used:
 
 Backwards-compatible changes increment the minor version number only.
 
+0.1.1
+-----
+
+Released: TBD
+
+* Reconnect to redis on errors (`#6 <https://github.com/sohonetlabs/nameko-rediskn/pull/6>`_)
+* New config key ``pubsub_backoff_factor`` (`#6 <https://github.com/sohonetlabs/nameko-rediskn/pull/6>`_)
 
 0.1.0
 -----